fix: updated colosseum multi boot data type [ci skip]

+++ src/intro_frlg.c
@@ -249,8 +249,8 @@ static void SpriteCB_NidorinoRecoil(struct Sprite *sprite);
 static void SpriteCB_NidorinoHop(struct Sprite *sprite);
 static void SpriteCB_NidorinoAttack(struct Sprite *sprite);
 
-extern const u32 gMultiBootProgram_PokemonColosseum_Start[];
-extern const u32 gMultiBootProgram_PokemonColosseum_End[];
+extern const u16 gMultiBootProgram_PokemonColosseum_Start[];
+extern const u16 gMultiBootProgram_PokemonColosseum_End[];
 
 static const u16 sCopyright_Pal[] = INCBIN_U16("graphics/intro_frlg/copyright.gbapal");
 static const u32 sCopyright_Gfx[]  = INCBIN_U32( "graphics/intro_frlg/copyright.4bpp.smol");
